Address review feedback, upgrade to 0.20.0
