--- parsers/test_kotlin_serialization_0_14_0/.gitignore
@@ -1,3 +1,0 @@
-*.iml
-target
-build
