--- parsers/test_kotlin_serialization_0_14_0/README.txt
@@ -1,3 +1,0 @@
-mvn clean package
-
-java -jar TestJSONParsing.jar
